32 bit data type for scan interval and scan window
The fields in the struct nrf_radio_request_earliest_t in nrf_soc.h are 32 bit long so even the values in btle.h have to be 32 bit long otherwise the behaviour is not correct. I have checked it looking at the power profile while scanning and noticed that the interval between scans was not correct.
I haven't checked for the advertising but I expect a similar problem.

+++ Include/btle.h
@@ -103,7 +103,7 @@ typedef enum
 typedef uint16_t btle_adv_interval_t;
 
 /** @brief BTLE scan interval type (_SHALL_ be fully compatible with the type in the HCI). */
-typedef uint16_t btle_scan_interval_t;
+typedef uint32_t btle_scan_interval_t;
 
 /** @brief BTLE connection interval type (_SHALL_ be fully compatible with the type in the HCI). */
 typedef uint16_t btle_conn_interval_t;
@@ -115,7 +115,7 @@ typedef uint16_t btle_connection_handle_t;
 typedef uint16_t btle_data_length_t;
 
 /** @brief BTLE scan window type (_SHALL_ be fully compatible with the type in the HCI). */
-typedef uint16_t btle_scan_window_t;
+typedef uint32_t btle_scan_window_t;
 
 /** @brief BTLE connection latency type (_SHALL_ be fully compatible with the type in the HCI). */
 typedef uint16_t  btle_conn_latency_t;
